case5/app/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-ulid
redis